refactor: 调整目录结构；使用gradle构建
