+++ src/main/java/com/qbb/component/ConfigComponent.java
