+++ src/main/java/com/qbb/component/ConfigPersistence.java
