+++ src/main/java/com/qbb/constant/JavaConstant.java
