+++ src/main/java/com/qbb/constant/ProjectTypeConstant.java
