+++ src/main/java/com/qbb/processor/AbstractClassProcessor.java
