+++ src/main/resources/META-INF/plugin.xml
@@ -1,36 +1,35 @@
 <idea-plugin>
 <id>YapiUpload</id>
 <name>YapiUpload</name>
-<version>1.8.2</version>
 <vendor email="chengsheng@qbb6.com" url="http://www.qbb6.com">QBB</vendor>
 
 <description><![CDATA[
-      upload api to yapi<br>
+      upload api to yapi<br/>
       <em>support dubbo api and api upload</em>
     ]]></description>
 
 <change-notes><![CDATA[
-      <a href="https://github.com/diwand/YapiIdeaUploadPlugin">how to use</a> <br>
-      1.8.2:<br> java type convert to json type <br>
-      1.8.1:<br> support idea 2020 <br>
-      1.7.9:<br>
+      <a href="https://github.com/diwand/YapiIdeaUploadPlugin">how to use</a> <br/>
+      1.8.2:<br/> java type convert to json type <br/>
+      1.8.1:<br/> support idea 2020 <br/>
+      1.7.9:<br/>
       <em>
       - fixbug
-     1.7.8:<br>
+     1.7.8:<br/>
      <em>
      - support method @path，fix remark bug
-     </em><br>
-     1.7.7:<br>
+     </em><br/>
+     1.7.7:<br/>
      <em>
       - support swagger & swagger mock
       - fixbug
-     </em><br>
-     1.7.6:<br>
+     </em><br/>
+     1.7.6:<br/>
       <em>
       - support base mock power
-      </em><br>
+      </em><br/>
 
-     1.7.5:<br>
+     1.7.5:<br/>
       <em>
       - Response complete support map
 
@@ -40,9 +39,9 @@
 
       - Support path missing / complete
 
-      </em><br>
+      </em><br/>
 
-      1.7.4:<br>
+      1.7.4:<br/>
 
       <em>
 
@@ -56,25 +55,25 @@
 
       - Unify multi module parameters to solve enumeration packing problem,
 
-      </em><br>
+      </em><br/>
 
-      1.7.3:<br>
+      1.7.3:<br/>
 
       <em>
 
       - Support map, solve the same package problem, remove cache, support notblank, support package menu
 
-      </em><br>
+      </em><br/>
 
-      1.7.2:<br>
+      1.7.2:<br/>
 
       <em>
 
       - New support for localdata, Localtime, localdatatime < br >
 
-      </em><br>
+      </em><br/>
 
-      1.7.1:<br>
+      1.7.1:<br/>
 
       <em>
 
@@ -86,9 +85,9 @@
 
       - Support multi-level classification (crazy branch)
 
-      </em><br>
+      </em><br/>
 
-      1.7.0:<br>
+      1.7.0:<br/>
 
       <em>
 
@@ -100,9 +99,9 @@
 
       - Link same package problem
 
-      </em><br>
+      </em><br/>
 
-      1.6.9:<br>
+      1.6.9:<br/>
 
       <em>
 
@@ -112,9 +111,9 @@
 
       - Support get entity class receiving
 
-      </em><br>
+      </em><br/>
 
-      1.6.8:<br>
+      1.6.8:<br/>
 
       <em>
 
@@ -126,7 +125,7 @@
 
       </em>
 
-      1.6.7:<br>
+      1.6.7:<br/>
 
       <em>
 
@@ -134,9 +133,9 @@
 
       - Support for pathvariable
 
-      <em> <br>
+      <em> <br/>
 
-      1.6.6:<br>
+      1.6.6:<br/>
 
       <em>
 
@@ -146,9 +145,9 @@
 
       - Support request param comments to get from comments
 
-      </em> <br>
+      </em> <br/>
 
-      1.6.5:<br>
+      1.6.5:<br/>
 
       <em>
 
@@ -160,9 +159,9 @@
 
       </em>
 
-      <br>
+      <br/>
 
-      1.6.4:<br>
+      1.6.4:<br/>
 
       <em>
 
@@ -170,9 +169,9 @@
 
       <em>
 
-      <br>
+      <br/>
 
-      1.6.3:<br>
+      1.6.3:<br/>
 
       <em>Fix multilevel generic problems
 
@@ -197,8 +196,6 @@
 -->
 
 <depends>com.intellij.modules.platform</depends>
-<depends>com.intellij.modules.lang</depends>
-<depends>com.intellij.modules.java</depends>
 
 <extensions defaultExtensionNs="com.intellij">
   <!-- Add your extensions here -->
@@ -219,7 +216,5 @@
     <keyboard-shortcut keymap="$default" first-keystroke="alt u"/>
   </action>
 </actions>
-
-<depends>com.intellij.modules.java</depends>
 
 </idea-plugin>
